+++ src/analysis/mod.rs
@@ -14,6 +14,7 @@ pub mod executor;
 pub mod objects;
 pub mod pass;
 pub mod rtti;
+pub mod seh;
 pub mod slices;
 pub mod tracker;
 pub mod vm;
